ReOrder ability (#118) (#122)
* merge fix

* remove /src/UI/Buyer/.gitIgnore

* remove karma.conf.js

* remove package-lock.json

* remove package.json

* remove readme

* these files were moved

* Feat: Reorder an Order

This is a container component which takes in a order id. validate the line items on the order, then display which items that are invalid and an add to cart button. Use Container because it calls out to the lineItemService to create line items with the valid products.

* Update reorder component to work with breaking changes

Unit test the order-reorder container

* refactor app lineitem service

* updated asked for changes

* Add Unit tests

update the service to throw an error if no idea is there.